+++ Summator/Summator.cs
@@ -22,7 +22,7 @@ namespace Summator
                 }
                 return sum;
             }
-        public static double Average(int[] arr)
+        public static double Average(double[] arr)
         {
             double sum = 0;
             for(int i=0; i<arr.Length; i++)
